Partial translation of Chapter 18 (up to line 156). Correction of typos in 'deteccionColision.py'.
