+++ translations/es/text/es_chapter18.figures.txt
@@ -1,0 +1,4 @@
+Fig 18-2:
+Intersecting Rectangles
+Not Intersecting Rectangles
+
